refactor: extract shared drag gesture utility and cleanup
- Remove unused 'steps' variable in AndroidDriver.drag()
- Extract duplicate drag implementation from WebDriver and CdpWebDriver
  to shared SeleniumGestures.performDrag() utility
- Add DRAG_LONG_PRESS_DURATION_MS constant (1000ms) for clarity
