+++ maestro-client/src/main/java/maestro/drivers/CdpWebDriver.kt
@@ -17,6 +17,7 @@ import maestro.SwipeDirection
 import maestro.TreeNode
 import maestro.ViewHierarchy
 import maestro.utils.ScreenshotUtils
+import maestro.utils.SeleniumGestures
 import maestro.web.record.JcodecVideoEncoder
 import maestro.web.record.WebScreenRecorder
 import okio.Sink
@@ -465,39 +466,7 @@ class CdpWebDriver(
 
     override fun drag(start: Point, end: Point, durationMs: Long) {
         val driver = ensureOpen()
-
-        val finger = PointerInput(PointerInput.Kind.TOUCH, "finger")
-        val drag = Sequence(finger, 1)
-        // Move to start position
-        drag.addAction(
-            finger.createPointerMove(
-                Duration.ofMillis(0),
-                PointerInput.Origin.viewport(),
-                start.x,
-                start.y
-            )
-        )
-        // Press and hold
-        drag.addAction(finger.createPointerDown(PointerInput.MouseButton.LEFT.asArg()))
-        // Pause to simulate long press before drag (1 second)
-        drag.addAction(finger.createPointerMove(
-            Duration.ofMillis(1000),
-            PointerInput.Origin.viewport(),
-            start.x,
-            start.y
-        ))
-        // Drag to end position
-        drag.addAction(
-            finger.createPointerMove(
-                Duration.ofMillis(durationMs - 1000),
-                PointerInput.Origin.viewport(),
-                end.x,
-                end.y
-            )
-        )
-        // Release
-        drag.addAction(finger.createPointerUp(PointerInput.MouseButton.LEFT.asArg()))
-        (driver as RemoteWebDriver).perform(listOf(drag))
+        SeleniumGestures.performDrag(driver as RemoteWebDriver, start, end, durationMs)
     }
 
     override fun backPress() {
